Project and Dissertation in Data Science - CSC8639 Leigh syndrome
Providing an accurate prognosis for patients suffering from Leigh syndrome is an unresolved challenge for clinicians and an active area of research. This is due to the huge number of variables that influence the progression of this disease. The aim of this project will be to build machine learning (ML) models that are appropriate for our data, can predict clinical outcomes for the patients and are interpretable.